nalgebra

+++ src/interval.rs
@@ -1,13 +1,13 @@
-use std::f64::{INFINITY, NEG_INFINITY};
+use std::f32::{INFINITY, NEG_INFINITY};
 
 #[derive(Debug, Clone, Copy)]
 pub struct Interval {
-    pub min: f64,
-    pub max: f64,
+    pub min: f32,
+    pub max: f32,
 }
 
 impl Interval {
-    pub fn new(min: f64, max: f64) -> Self {
+    pub fn new(min: f32, max: f32) -> Self {
         Self { min, max }
     }
 
@@ -18,19 +18,19 @@ impl Interval {
         }
     }
 
-    pub fn size(&self) -> f64 {
+    pub fn size(&self) -> f32 {
         self.max - self.min
     }
 
-    pub fn contains(&self, x: f64) -> bool {
+    pub fn contains(&self, x: f32) -> bool {
         self.min <= x && x <= self.max
     }
 
-    pub fn surrounds(&self, x: f64) -> bool {
+    pub fn surrounds(&self, x: f32) -> bool {
         self.min < x && x < self.max
     }
 
-    pub fn clamp(&self, x: f64) -> f64 {
+    pub fn clamp(&self, x: f32) -> f32 {
         if (x < self.min) {
             return self.min
         }
@@ -42,7 +42,7 @@ impl Interval {
     }
 
     // pads an interval so no NaN
-    pub fn expand(&self, delta: f64) -> Interval {
+    pub fn expand(&self, delta: f32) -> Interval {
         let padding = delta/2.;
         Interval::new(self.min - padding, self.max + padding)
     }
